:truck: (Information) move component from 'search/' to 'components/'
