+++ frontend/src/components/Information/BaseInformation.tsx
