+++ frontend/src/components/Information/RemoteIconInformation/index.tsx
